+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/Reactions/MessageReactionsCellDescription.swift
@@ -19,8 +19,6 @@
 import UIKit
 import WireDataModel
 
-// MARK: - MessageReactionsCellDescription
-
 final class MessageReactionsCellDescription: ConversationMessageCellDescription {
 
     // MARK: - Properties
@@ -31,8 +29,10 @@ final class MessageReactionsCellDescription: ConversationMessageCellDescription 
     init(message: ZMConversationMessage) {
         self.message = message
         configuration = message.reactionsSortedByCreationDate().compactMap { reaction in
+
             guard !reaction.users.isEmpty else { return nil }
-            return MessageReactionMetadata(
+
+            return MessageReaction(
                 emojiID: reaction.reactionString,
                 count: UInt(reaction.users.count),
                 isSelfUserReacting: reaction.users.contains(where: \.isSelfUser)
